+++ src/matchers/Emoji.js
@@ -14,7 +14,7 @@ import {
   UNICODE_TO_SHORTNAME,
 } from '../data/emoji';
 
-import type { MatchResponse, EmojiProps, EmojiOptions } from '../types';
+import type { MatchResponse, EmojiProps, EmojiOptions, ParsedNodes } from '../types';
 
 const EMOJI_REGEX = new RegExp(EMOJI_PATTERN);
 const EMOJI_SHORTNAME_REGEX = new RegExp(EMOJI_SHORTNAME_PATTERN, 'i');
@@ -83,5 +83,28 @@ export default class EmojiMatcher extends Matcher<EmojiOptions> {
     }
 
     return response;
+  }
+
+  /**
+   * {@inheritDoc}
+   */
+  onAfterParse(content: ParsedNodes): ParsedNodes {
+    // When a single `Emoji` is the only content, enlarge it!
+    if (content.length === 1) {
+      let item = content[0];
+
+      if (typeof item !== 'string' && React.isValidElement(item) && item.type === Emoji) {
+        item = (item: React.Element<*>);
+
+        return [
+          React.cloneElement(item, {
+            ...item.props,
+            enlargeEmoji: true,
+          }),
+        ];
+      }
+    }
+
+    return content;
   }
 }